Match and sort columns with regex capture groups

Match and sort columns with regex capture groups

Usage

match_order(match, ordering, ..., order_by = c("fastest", "slowest"))

Arguments

match

A regular expression. Passed to tidyselect::matches().

ordering

An integer vector of regex capture group IDs to order the columns by. Negative indices sort the group in reverse order. If not specified, defaults to the mutually exclusive argument order_by.

...

Arguments passed on to tidyselect::matches

ignore.case

If TRUE, the default, ignores case when matching names.

vars

A character vector of variable names. If not supplied, the variables are taken from the current selection context (as established by functions like select() or pivot_longer()).

perl

Should Perl-compatible regexps be used?

order_by

One of "fastest" or "slowest". Determines the speed of cycling through column groups. Mutually exclusive with ordering. Defaults to "fastest".

See also

Examples

library(dplyr, warn.conflicts = FALSE)
df <- data.frame(
  x_1 = 1,
  x_2 = 2,
  y_2 = 3,
  y_1 = 4,
  y_3 = 5,
  x_3 = 6
)

# Comparison: `matches()`
df %>%
  select(matches("(x|y)_(\\d)"))
#>   x_1 x_2 y_2 y_1 y_3 x_3
#> 1   1   2   3   4   5   6

# Order with priority on keeping the letters (1st group) close together
df %>%
  select(match_order("(x|y)_(\\d)", c(1, 2)))
#>   x_1 x_2 x_3 y_1 y_2 y_3
#> 1   1   2   6   4   3   5
# Same as above, but sort letters in reverse order
df %>%
  select(match_order("(x|y)_(\\d)", c(-1, 2)))
#>   y_1 y_2 y_3 x_1 x_2 x_3
#> 1   4   3   5   1   2   6

# Order with priority on keeping the numbers (2nd group) close together
df %>%
  select(match_order("(x|y)_(\\d)", c(2, 1)))
#>   x_1 y_1 x_2 y_2 x_3 y_3
#> 1   1   4   2   3   6   5
# Same as above, the 2nd group is "slowest" because it has more categories
df %>%
  select(match_order("(x|y)_(\\d)", order_by = "slowest"))
#>   x_1 x_2 x_3 y_1 y_2 y_3
#> 1   1   2   6   4   3   5